Extract DocumentSideAcuantCapture component
**Why**: Better assurances that behaviors are kept consistent between DocumentsStep and ReviewIssuesStep
